modified:   camera_imu_syncer/camera_imu_syncer_node/include/camera_imu_syncer_node/Camera.hpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/include/camera_imu_syncer_node/Imu.hpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/include/camera_imu_syncer_node/Syncer.hpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/lib/MVSDK/MindVisionCamera.cpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/lib/MVSDK/MindVisionCamera.h 	modified:   camera_imu_syncer/camera_imu_syncer_node/src/Camera.cpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/src/Imu.cpp 	modified:   camera_imu_syncer/camera_imu_syncer_node/src/Syncer.cpp

+++ camera_imu_syncer/camera_imu_syncer_node/include/camera_imu_syncer_node/Camera.hpp
@@ -7,7 +7,7 @@ private:
     std::shared_ptr<Syncer> syncer;
 
     MindVisionCamera camera;
-    std::function<void(tSdkFrameHead* pHead, BYTE* pBuffer, int timeDiff)> frameCallback;
+    std::function<void(tSdkFrameHead* pHead, BYTE* pBuffer, int64_t timeDiff)> frameCallback;
 
 public:
     std::mutex mtx;
